% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs2dlm.R
\name{pbs2dlm}
\alias{pbs2dlm}
\title{Convert groundfish PBS data to a DLMtool data object}
\usage{
pbs2dlm(dat, name = "", area = "3[CD]+", survey = "SYN WCVI",
  max_year = 2017, min_mean_length = 10, length_bin_interval = 2)
}
\arguments{
\item{dat}{A list object output from \code{\link[gfplot:cache_pbs_data]{gfplot::cache_pbs_data()}} or a list
object containing the elements \code{commercial_samples}, \code{survey_samples},
\code{catch}, \code{survey_index}.}

\item{name}{A name for the stock.}

\item{area}{The groundfish statistical area to subset the catch by.}

\item{survey}{A survey abbreviation designating which survey to use for the
relative index of abundance.}

\item{max_year}{The most recent year of data to include.}

\item{min_mean_length}{The minimum number of samples to include a mean length
measurement for a given year.}

\item{length_bin_interval}{An interval for the length bins.}
}
\value{
An S4 object of class DLMtool Data.
}
\description{
Takes the output from \pkg{gfplot} data fetching functions and converts them to a
DLMtool data object.
}
\examples{
\dontrun{
library(gfplot)
species <- "pacific cod"
d <- list()
d$commercial_samples <- get_commercial_samples(species)
d$survey_samples <- get_survey_samples(species)
d$catch <- get_catch(species)
d$survey_index <- get_survey_index(species)
pbs2dlm(d, name = "BC Pacific Cod", area = "3[CD]+")
}
}
